arturo/app/requirements/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-crum==0.7.5
django-widget-tweaks==1.4.5
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1

